feat(TextInput): Replace deprecated implementation
