+++ packages/react-component-library/cypress/selectors/form.ts
@@ -1,7 +1,7 @@
 export default {
   input: {
-    email: '[data-testid="form-example-TextInputE-email"]',
-    password: '[data-testid="form-example-TextInputE-password"]',
+    email: '[data-testid="form-example-TextInput-email"]',
+    password: '[data-testid="form-example-TextInput-password"]',
     description: '[data-testid="form-example-TextArea-description"]',
     checkbox: '[data-testid="checkbox"]',
     radio: '[data-testid="radio"]',
