+++ packages/react-component-library/src/components/DatePicker/DatePicker.tsx
@@ -13,12 +13,12 @@ import { hasClass } from '../../helpers'
 import { InlineButton } from '../InlineButtons/InlineButton'
 import { InputValidationProps } from '../../common/InputValidationProps'
 import { isDateValid } from './utils'
-import { StyledLabel } from '../TextInputE/partials/StyledLabel'
+import { StyledLabel } from '../TextInput/partials/StyledLabel'
 import { StyledDatePickerInput } from './partials/StyledDatePickerInput'
 import { StyledDayPicker } from './partials/StyledDayPicker'
 import { StyledFloatingBox } from './partials/StyledFloatingBox'
 import { StyledInlineButtons } from '../InlineButtons/partials/StyledInlineButtons'
-import { StyledInput } from '../TextInputE/partials/StyledInput'
+import { StyledInput } from '../TextInput/partials/StyledInput'
 import { StyledInputWrapper } from './partials/StyledInputWrapper'
 import { StyledOuterWrapper } from './partials/StyledOuterWrapper'
 import { useExternalId } from '../../hooks/useExternalId'
