+++ packages/react-component-library/src/components/SelectBase/SelectLayout.tsx
@@ -10,7 +10,7 @@ import { SelectChildWithStringType } from './types'
 import { StyledInlineButtons } from './partials/StyledInlineButtons'
 import { StyledInput } from './partials/StyledInput'
 import { StyledInputWrapper } from './partials/StyledInputWrapper'
-import { StyledLabel } from '../TextInputE/partials/StyledLabel'
+import { StyledLabel } from '../TextInput/partials/StyledLabel'
 import { StyledOptions } from './partials/StyledOptions'
 import { StyledOptionsWrapper } from './partials/StyledOptionsWrapper'
 import { StyledOuterWrapper } from './partials/StyledOuterWrapper'
