+++ packages/react-component-library/src/components/SelectBase/partials/StyledTextInput.tsx
@@ -1,6 +1,6 @@
 import styled from 'styled-components'
 
-import { StyledTextInput as StyledTextInputBase } from '../../TextInputE/partials/StyledTextInput'
+import { StyledTextInput as StyledTextInputBase } from '../../TextInput/partials/StyledTextInput'
 
 export const StyledTextInput = styled(StyledTextInputBase)`
   margin-bottom: 0;
