+++ packages/react-component-library/src/components/TextInput/TextInput.stories.tsx
@@ -7,7 +7,9 @@ import { TextInput } from '.'
 export default {
   component: TextInput,
   title: 'Text Input',
-  argTypes: { onBlur: { action: 'onBlur' } },
+  parameters: {
+    argTypes: { onBlur: { action: 'onBlur' } },
+  },
 } as ComponentMeta<typeof TextInput>
 
 export const Default: ComponentStory<typeof TextInput> = (props) => (
@@ -18,12 +20,6 @@ Default.args = {
   name: 'text-input-default',
   label: 'Example label',
 }
-
-export const WithLabel: ComponentStory<typeof TextInput> = (props) => (
-  <TextInput {...props} name="text-input-label" label="Example label" />
-)
-
-WithLabel.storyName = 'With label'
 
 export const Disabled: ComponentStory<typeof TextInput> = (props) => (
   <TextInput
@@ -57,3 +53,14 @@ export const WithEndAdornment: ComponentStory<typeof TextInput> = (props) => (
 )
 
 WithEndAdornment.storyName = 'With end adornment'
+
+export const WithError: ComponentStory<typeof TextInput> = (props) => (
+  <TextInput
+    {...props}
+    isInvalid
+    label="Example label"
+    name="text-input-end-adornment"
+  />
+)
+
+WithError.storyName = 'With error'
