--- packages/react-component-library/src/components/TextInputE/index.ts
@@ -1,1 +1,0 @@
-export * from './TextInputE'
