--- packages/react-component-library/src/components/TextInputE/partials/StyledInputWrapper.tsx
@@ -1,7 +1,0 @@
-import styled from 'styled-components'
-
-export const StyledInputWrapper = styled.div`
-  position: relative;
-  flex-grow: 1;
-  order: 2;
-`
