--- packages/react-component-library/src/components/TextInputE/partials/StyledOuterWrapper.tsx
@@ -1,13 +1,0 @@
-import styled from 'styled-components'
-
-import {
-  StyledOuterWrapper as StyledOuterWrapperBase,
-  StyledOuterWrapperProps,
-} from '../../../styled-components'
-
-export const StyledOuterWrapper = styled(
-  StyledOuterWrapperBase
-)<StyledOuterWrapperProps>`
-  display: inline-flex;
-  flex-direction: row;
-`
