+++ packages/react-component-library/src/components/TopLevelNavigation/Masthead/Masthead.test.tsx
@@ -124,7 +124,6 @@ describe('Masthead', () => {
           </Notifications>
         ),
         onSearch: () => true,
-        searchPlaceholder: 2,
         title: 'title',
         user: (
           <MastheadUser
