+++ packages/react-component-library/src/components/TopLevelNavigation/SearchBar/SearchBar.tsx
@@ -11,7 +11,6 @@ export interface SearchbarProps {
   className?: string
   onSearch: (event: React.FormEvent<HTMLFormElement>, term: string) => void
   searchButton: any
-  searchPlaceholder: string
   setShowSearch: (isVisible: boolean) => void
   style?: Record<string, unknown>
 }
@@ -19,7 +18,6 @@ export interface SearchbarProps {
 export const SearchBar: React.FC<SearchbarProps> = ({
   onSearch,
   searchButton,
-  searchPlaceholder,
   setShowSearch,
   ...rest
 }) => {
@@ -49,10 +47,10 @@ export const SearchBar: React.FC<SearchbarProps> = ({
           autoFocus
           id="term"
           name="term"
-          onChange={(event) => {
+          label=""
+          onChange={(event: React.ChangeEvent<HTMLInputElement>) => {
             setTerm(event.target.value)
           }}
-          placeholder={searchPlaceholder}
           value={term}
         />
         <StyledButton
