+++ packages/react-component-library/src/index.ts
@@ -47,7 +47,6 @@ export * from './components/TabNav'
 export * from './components/TabSet'
 export * from './components/TextArea'
 export * from './components/TextInput'
-export * from './components/TextInputE'
 export * from './components/Timeline'
 export * from './components/Toast'
 export * from './components/Tooltip'
